+++ src/tools/__init__.py
@@ -4,3 +4,8 @@ Public tool functions live in domain modules. ``src.tool_implementations``
 re-exports from here for backward compatibility.
 """
 from src.tools._common import _parse_tool_args  # noqa: F401
+from src.tools.system import (  # noqa: F401
+    do_manage_skills, _skill_dump, do_manage_tasks, do_manage_endpoints,
+    do_manage_mcp, do_manage_webhooks, do_manage_tokens, do_manage_settings,
+    do_api_call, do_app_api,
+)
